docs, move jsonpath, errors, metadata
