+++ miniconf/tests/generics.rs
@@ -37,9 +37,9 @@ fn generic_array() {
     assert_eq!(settings.data[0], 3.0);
 
     // Test metadata
-    let metadata = Settings::<f32>::metadata().separator("/");
+    let metadata = Settings::<f32>::metadata();
     assert_eq!(metadata.max_depth, 2);
-    assert_eq!(metadata.max_length, "/data/0".len());
+    assert_eq!(metadata.max_length("/"), "/data/0".len());
     assert_eq!(metadata.count, 2);
 }
 
@@ -61,9 +61,9 @@ fn generic_struct() {
     assert_eq!(settings.inner.data, 3.0);
 
     // Test metadata
-    let metadata = Settings::<Inner>::metadata().separator("/");
+    let metadata = Settings::<Inner>::metadata();
     assert_eq!(metadata.max_depth, 1);
-    assert_eq!(metadata.max_length, "/inner".len());
+    assert_eq!(metadata.max_length("/"), "/inner".len());
     assert_eq!(metadata.count, 1);
 }
 
@@ -91,9 +91,9 @@ fn generic_atomic() {
     assert_eq!(settings.atomic.inner[0], 3.0);
 
     // Test metadata
-    let metadata = Settings::<f32>::metadata().separator("/");
+    let metadata = Settings::<f32>::metadata();
     assert_eq!(metadata.max_depth, 3);
-    assert_eq!(metadata.max_length, "/opt1/0/0".len());
+    assert_eq!(metadata.max_length("/"), "/opt1/0/0".len());
 }
 
 #[test]
